refactor deleting sensor data
