+++ lib/models/box.js
@@ -24,7 +24,7 @@ const { mongoose } = require('../db'),
   integrations = require('./integrations'),
   sensorLayouts = require('../sensorLayouts'),
   mqttClient = require('../mqtt'),
-  Measurement = require('./measurement').model,
+  { model: Measurement } = require('./measurement'),
   {
     config: { api_measurements_post_domain, imageFolder },
     parseTimestamp,
@@ -315,36 +315,94 @@ boxSchema.statics.findBoxById = function findBoxById (id, { lean = true, populat
     });
 };
 
-boxSchema.statics.findByIdAndUpdateLastMeasurementOfSensor = function findByIdAndUpdateLastMeasurementOfSensor (id, sensorId, shouldQueryLatestMeasurement) {
-  let newLastMeasurementPromise = Promise.resolve();
-  if (typeof shouldQueryLatestMeasurement !== 'undefined' && shouldQueryLatestMeasurement === true) {
-    newLastMeasurementPromise = Measurement.findLastMeasurementOfSensor(sensorId);
+const DELETE_MEASUREMENTS_CONTACT_ADMIN_MSG = 'If you feel your box may be in inconsistent state, please contact the administrator.';
+const DELETE_MEASUREMENTS_UNSUCCESSFUL_ERROR = 'Delete operation partially unsuccessful. This usually means some criteria you specified didn\'t yield measurements to delete or the sensor had no measurements. This can happen, if you send the same request twice.';
+
+boxSchema.methods.deleteMeasurementsOfSensor = function deleteMeasurementsOfSensor ({ sensorId, deleteAllMeasurements, timestamps, fromDate, toDate }) {
+  const box = this;
+
+  const sensorIndex = box.sensors.findIndex(s => s._id.equals(sensorId));
+  if (sensorIndex === -1) {
+    throw new ModelError(`Sensor with id ${sensorId} not found.`, { type: 'NotFoundError' });
   }
 
-  return newLastMeasurementPromise
-    .then((newLastMeasurement) => { // arrow function for scope
-      if (newLastMeasurement && Array.isArray(newLastMeasurement) && newLastMeasurement.length !== 0) {
-        newLastMeasurement = newLastMeasurement[0]._id;
-      } else { // reply was just an empty array
-        newLastMeasurement = undefined;
+  const reallyDeleteAllMeasurements = (deleteAllMeasurements === 'true');
+  // check for instruction exclusivity
+  if (reallyDeleteAllMeasurements === true && (timestamps || (fromDate && toDate))) {
+    return Promise.reject(new ModelError('Parameter deleteAllMeasurements can only be used by itself'));
+  } else if (reallyDeleteAllMeasurements === false && timestamps && fromDate && toDate) {
+    return Promise.reject(new ModelError('Please specify only timestamps or a range with from-date and to-date'));
+  } else if (reallyDeleteAllMeasurements === false && !timestamps && !fromDate && !toDate) {
+    return Promise.reject(new ModelError('DeleteAllMeasurements not true. deleting nothing'));
+  }
+
+  let successMsg = 'all measurements',
+    mode = 'all';
+  let createdAt;
+
+  if (timestamps) {
+    createdAt = {
+      $in: timestamps.map(t => t.toDate())
+    };
+    successMsg = `${timestamps.length} measurements`;
+    mode = 'timestamps';
+  }
+
+  if (fromDate && toDate) {
+    createdAt = {
+      $gt: fromDate.toDate(),
+      $lt: toDate.toDate()
+    };
+    successMsg = `measurements between ${fromDate.format()} and ${toDate.format()}`;
+    mode = 'range';
+  }
+
+  const query = {
+    sensor_id: sensorId
+  };
+
+  if (createdAt) {
+    query.createdAt = createdAt;
+  }
+
+  let deleteUnsuccessful = false;
+
+  // delete measurements
+  return Measurement.find(query)
+    .remove()
+    .exec()
+    .then(function (removeResult) {
+      if (removeResult && removeResult.result && removeResult.result.n === 0) {
+        throw new ModelError('No matching measurements for specified query', { type: 'NotFoundError' });
+      }
+      // check for not ok deletion, this is not a failure but should generate a warning for the user
+      if (removeResult.result.ok !== 1) {
+        deleteUnsuccessful = true;
       }
 
-      return this.findById(id) // scope is used here
-        .populate('sensor.lastMeasurement')
-        .exec()
-        .then(function (box) {
-          const sensorIndex = box.sensors.findIndex(s => s._id.equals(sensorId));
-          if (sensorIndex === -1) {
-            throw new Error('SENSOR_NOT_FOUND');
-          }
+      let newLastMeasurementPromise = Promise.resolve();
+      if (mode !== 'all') {
+        newLastMeasurementPromise = Measurement.findLastMeasurementOfSensor(sensorId);
+      }
 
+      return newLastMeasurementPromise
+        .then(function (newLastMeasurement) {
           box.set(`sensors.${sensorIndex}.lastMeasurement`, newLastMeasurement);
 
           return box.save();
         });
-    });
-};
+    })
+    .then(function () {
+      let responseMessage = `Successfully deleted ${successMsg} of sensor ${sensorId}`;
 
+      if (deleteUnsuccessful === true) {
+        responseMessage = `${DELETE_MEASUREMENTS_UNSUCCESSFUL_ERROR} ${DELETE_MEASUREMENTS_CONTACT_ADMIN_MSG}`;
+      }
+
+      return responseMessage;
+    });
+
+};
 
 /**
  * updates a boxes location at a given time, and performs housekeeping logic.
