+++ lib/models/measurement.js
@@ -53,40 +53,16 @@ measurementSchema.set('toJSON', {
   }
 });
 
-measurementSchema.statics.deleteMeasurementsOfSensor = function deleteMeasurementsOfSensor (sensorId, createdAt) {
-  const query = {
-    sensor_id: sensorId
-  };
-
-  if (createdAt) {
-    query.createdAt = createdAt;
-  }
-
-  // delete measurements
-  return this.find(query)
-    .remove()
-    .exec()
-    .then(function (removeResult) {
-      const errors = [];
-      // nothing removed -> lastMeasurement should not be updated
-      if (removeResult && removeResult.result && removeResult.result.n === 0) {
-        errors.push('NO_MATCHING_MEASUREMENTS');
-        //throw new BadRequestError('no matching measurements for specified query');
-      }
-      // check for not ok deletion, this is not a failure but should generate a warning for the user
-      if (removeResult.result.ok !== 1) {
-        errors.push('DELETE_ERROR');
-      }
-
-      return errors;
-    });
-};
-
 measurementSchema.statics.findLastMeasurementOfSensor = function findLastMeasurementOfSensor (sensorId) {
   return this.find({ sensor_id: sensorId })
     .sort({ createdAt: -1 })
     .limit(1)
-    .exec();
+    .exec()
+    .then(function (lastMeasurement) {
+      if (lastMeasurement && Array.isArray(lastMeasurement) && lastMeasurement.length !== 0) {
+        return lastMeasurement[0];
+      }
+    });
 };
 
 measurementSchema.statics.decodeMeasurements = function decodeMeasurements (measurements, { contentType = 'json', sensors }) {
